Updated README. Added file check safety

+++ core/src/main/java/de/tum/cit/fop/maze/level/GameOverScreen.java
@@ -185,9 +185,15 @@ public class GameOverScreen implements Screen {
      */
     public void deleteGame() {
         int index = LevelScreen.getInstance().getLevelIndex();
-        Gdx.files.local("saves/" + index + ".json").delete();
-        Gdx.files.local("saves/" + index + ".png").delete();
-        Gdx.files.local("saves/levelData_" + index + ".json").delete();
+        if (Gdx.files.local("saves/" + index + ".json").exists()) {
+            Gdx.files.local("saves/" + index + ".json").delete();
+        }
+        if (Gdx.files.local("saves/" + index + ".png").exists()){
+            Gdx.files.local("saves/" + index + ".png").delete();
+        }
+        if (Gdx.files.local("saves/levelData_" + index + ".json").exists()){
+            Gdx.files.local("saves/levelData_" + index + ".json").delete();
+        }
     }
 
     @Override
